Only launch FlashActivity if app is foreground
