+++ app/src/main/java/com/topjohnwu/magisk/App.kt
@@ -1,13 +1,11 @@
 package com.topjohnwu.magisk
 
 import android.annotation.SuppressLint
-import android.app.Activity
 import android.app.Application
 import android.content.Context
 import android.content.res.Configuration
 import android.os.AsyncTask
 import android.os.Build
-import android.os.Bundle
 import androidx.appcompat.app.AppCompatDelegate
 import androidx.multidex.MultiDex
 import androidx.room.Room
@@ -16,7 +14,7 @@ import androidx.work.impl.WorkDatabase_Impl
 import com.topjohnwu.magisk.data.database.RepoDatabase
 import com.topjohnwu.magisk.data.database.RepoDatabase_Impl
 import com.topjohnwu.magisk.di.koinModules
-import com.topjohnwu.magisk.extensions.inject
+import com.topjohnwu.magisk.extensions.get
 import com.topjohnwu.magisk.utils.LocaleManager
 import com.topjohnwu.magisk.utils.RootUtils
 import com.topjohnwu.net.Networking
@@ -26,12 +24,9 @@ import org.koin.core.context.startKoin
 import timber.log.Timber
 import java.util.concurrent.ThreadPoolExecutor
 
-open class App : Application(), Application.ActivityLifecycleCallbacks {
+open class App : Application() {
 
     lateinit var protectedContext: Context
-
-    @Volatile
-    private var foreground: Activity? = null
 
     override fun attachBaseContext(base: Context) {
         super.attachBaseContext(base)
@@ -54,7 +49,7 @@ open class App : Application(), Application.ActivityLifecycleCallbacks {
             deContext.moveSharedPreferencesFrom(base, base.defaultPrefsName)
         }
 
-        registerActivityLifecycleCallbacks(this)
+        registerActivityLifecycleCallbacks(get())
 
         Networking.init(base)
         LocaleManager.setLocale(this)
@@ -64,28 +59,6 @@ open class App : Application(), Application.ActivityLifecycleCallbacks {
         super.onConfigurationChanged(newConfig)
         LocaleManager.setLocale(this)
     }
-
-    //region ActivityLifecycleCallbacks
-    override fun onActivityCreated(activity: Activity, bundle: Bundle?) {}
-
-    override fun onActivityStarted(activity: Activity) {}
-
-    @Synchronized
-    override fun onActivityResumed(activity: Activity) {
-        foreground = activity
-    }
-
-    @Synchronized
-    override fun onActivityPaused(activity: Activity) {
-        foreground = null
-    }
-
-    override fun onActivityStopped(activity: Activity) {}
-
-    override fun onActivitySaveInstanceState(activity: Activity, bundle: Bundle) {}
-
-    override fun onActivityDestroyed(activity: Activity) {}
-    //endregion
 
     private val Context.defaultPrefsName get() = "${packageName}_preferences"
 
@@ -119,13 +92,6 @@ open class App : Application(), Application.ActivityLifecycleCallbacks {
                     else -> null
                 }
             }
-        }
-
-        @Deprecated("")
-        @JvmStatic
-        fun foreground(): Activity? {
-            val app: App by inject()
-            return app.foreground
         }
     }
 }
