+++ app/src/main/java/com/topjohnwu/magisk/extensions/XKoin.kt
@@ -6,12 +6,12 @@ import org.koin.core.qualifier.Qualifier
 
 fun getKoin() = GlobalContext.get().koin
 
-inline fun <reified T : Any> inject(
+inline fun <reified T> inject(
     qualifier: Qualifier? = null,
     noinline parameters: ParametersDefinition? = null
 ) = lazy { get<T>(qualifier, parameters) }
 
-inline fun <reified T : Any> get(
+inline fun <reified T> get(
     qualifier: Qualifier? = null,
     noinline parameters: ParametersDefinition? = null
 ): T = getKoin().get(qualifier, parameters)
